+++ src/main/scala/models/AggregateDetails.scala
@@ -2,5 +2,5 @@ package models
 
 case class AggregateDetails(otherTotalWealth: Long,
                             oecdTotalWealth: Long,
-                            otherTotalPopulation: Int,
-                            oecdTotalPopulation: Int)
+                            otherTotalPopulation: Long,
+                            oecdTotalPopulation: Long)
