+++ src/main/scala/models/CountryPopulationDetail.scala
@@ -1,5 +1,5 @@
 package models
 
 case class CountryPopulationDetail(name: String,
-                                   population: Int
+                                   population: Long
                         )
